Debugged routing from recent pull and made adjustments to nav-bar
Fixed bugs in the routing for pages that weren't working that resulted from pulling changes from feat-HomePage branch. Added affix property to the entire navbar to make it stay in place while scrolling and added some CSS to give the bottom of the nav-bar a subtle drop shadow for better separation.