Image view modal
Adjusted Z-index to modal in AlbumPage for "Add New Album" feature to make it appear on top of the nav bar and not behind it. In open-album page, adjusted the size of the thumbnails to 50px X 50px to align with how the backend will process thumbnails. Added onclick modal feature to thumbnails to see a full size view of the thumbnail images.