Debugged feature to add an image to an album
The post("/api/albums/{id}") enpoint to add an image to an album was returning 200 in the body response, but it wasn't updating the database with the new image. Debugged the update query in the endpoint that adds an image to an album. Changed it from using the string version of the ObjectId to converting it from a string to the ObjectId and passing the ObjectId to MongoDb.